Refactor API code into services package
Handlers are now only responsible for the handling of the HTTP request and response. Reading and writing to the datastore, and other logic for ensuring we don't have videostreams pointing to non-existent capture sources etc, has been moved into services. This will make code easier to understand, and make unit testing easier.

+++ api/model/annotations.go
@@ -61,7 +61,7 @@ type BoundingBox struct {
 
 // An Annotation holds information about observations at a particular moment and region within a video stream.
 type Annotation struct {
-	VideoStreamID    int
+	VideoStreamID    int64
 	TimeSpan         TimeSpan
 	BoundingBox      *BoundingBox // Optional.
 	Observer         string
